requirements.txt
Django==4.2
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
pillow==11.1.0
